+++ src/storage/block_compactor.cpp
@@ -325,7 +325,7 @@ void BlockCompactor::BuildDictionary(std::vector<const byte *> *loose_ptrs, Arro
   }
   ArrowColumnInfo new_col_info;
   new_col_info.Type() = col->Type();
-  new_col_info.Indices() = common::AllocationUtil::AllocateAligned<uint32_t>(metadata->NumRecords());
+  new_col_info.Indices() = common::AllocationUtil::AllocateAligned<uint64_t>(metadata->NumRecords());
   auto &new_col = new_col_info.VarlenColumn() = {varlen_size, static_cast<uint32_t>(dictionary.size() + 1)};
 
   // TODO(Tianyu): This is retarded, but apparently you cannot retrieve the index of elements in your
@@ -352,7 +352,7 @@ void BlockCompactor::BuildDictionary(std::vector<const byte *> *loose_ptrs, Arro
     VarlenEntry &entry = values[i];
     // Need to GC
     if (entry.NeedReclaim()) loose_ptrs->push_back(entry.Content());
-    uint32_t dictionary_code = new_col_info.Indices()[i] = dictionary[entry];
+    uint64_t dictionary_code = new_col_info.Indices()[i] = dictionary[entry];
 
     byte *dictionary_word = new_col.Values() + new_col.Offsets()[dictionary_code];
     TERRIER_ASSERT(memcmp(dictionary_word, entry.Content(), entry.Size()) == 0,
